requirements.txt
aiosqlite~=0.20.0
python-dotenv~=1.0.1
redis~=5.1.1
aiogram~=3.13.1
SQLAlchemy~=2.0.36
PyYAML~=6.0.2
alembic~=1.14.0